проблема с обновлением старых точек
