+++ package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "dayjs": "1.11.1",
-    "flatpickr": "4.6.13"
+    "flatpickr": "4.6.13",
+    "he": "1.2.0"
   }
 }
