+++ src/main.js
@@ -2,9 +2,9 @@ import BoardPresenter from './presenter/board-presenter.js';
 import FilterPresenter from './presenter/filter-presenter.js';
 import InfoPresenter from './presenter/info-presenter.js';
 import PointModel from './model/point-model.js';
-import OffersModel from './model/offers-model.js';
+// import OffersModel from './model/offers-model.js';
 import FilterModel from './model/filter-model.js';
-import DestinationsModel from './model/destinations-model.js';
+// import DestinationsModel from './model/destinations-model.js';
 import PointsApiService from './api-services/points-api-service.js';
 import DestinationsApiService from './api-services/destinations-api-service.js';
 import OffersApiService from './api-services/offers-api-services.js';
@@ -22,17 +22,15 @@ const siteTripControlsFilters = siteHeaderElement.querySelector('.trip-controls_
 const newPointButton = document.querySelector('.trip-main__event-add-btn');
 
 
-const pointModel = new PointModel(new PointsApiService(END_POINT, AUTHORIZATION));
-const destinationsModel = new DestinationsModel(new DestinationsApiService(END_POINT, AUTHORIZATION));
-const offersModel = new OffersModel(new OffersApiService(END_POINT, AUTHORIZATION));
+const pointModel = new PointModel(new PointsApiService(END_POINT, AUTHORIZATION), new DestinationsApiService(END_POINT, AUTHORIZATION), new OffersApiService(END_POINT, AUTHORIZATION) );
 
 
 const filterModel = new FilterModel();
 
 
 const filterPresenter = new FilterPresenter(siteTripControlsFilters, filterModel, pointModel);
-const boardPresenter = new  BoardPresenter(sitePointsElement, siteHeaderMainElement, pointModel, offersModel, filterModel, destinationsModel);
-const infoPresenter = new InfoPresenter(siteHeaderMainElement, filterModel, pointModel, offersModel);
+const boardPresenter = new  BoardPresenter(sitePointsElement, siteHeaderMainElement, pointModel, filterModel);
+const infoPresenter = new InfoPresenter(siteHeaderMainElement, filterModel, pointModel);
 
 const handleNewPointFormClose = () => {
   newPointButton.disabled = false;
@@ -43,8 +41,7 @@ const handleNewPointButtonClick = () =>{
   newPointButton.disabled = true;
 };
 
-offersModel.init();
-destinationsModel.init();
+
 pointModel.init()
   .finally(()=>{
     newPointButton.disabled = false;
